Kcal-stats : update detail description for fitness tracking info -improve graph value

+++ include/apps/tools/OswAppKcalStats.h
@@ -20,7 +20,7 @@ class OswAppKcalStats : public OswApp {
   void showCurvedChart();
   void movingValue();
   void readyValue();
-  int weekValue[7];
+  uint16_t weekValue[7];
   OswUI* ui;
 };
